feat: (WIP) set up subroutine files (audit_logs.py, tenants.py) for separate classes, implement APIRouter and db.py to separate endpoints according to classes
