+++ routers/__init__.py
@@ -1,0 +1,4 @@
+from .audit_logs import router as audit_logs_router
+from .tenants    import router as tenants_router
+
+__all__ = ["audit_logs_router", "tenants_router"]
